natural4-server/pyproject.toml
[project]
name = "natural4-server"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = "==3.12.*"


dependencies = [
    "aioshutil == 1.5",
    "aiostream == 0.5.2",
    "anyio == 4.6.0",
    "cytoolz == 0.12.3",
    "jinja2 == 3.1.4",
    "orjson == 3.10.7",
    "pypandoc-binary == 1.13",
    "pyrsistent == 0.20.0",
    "pyrsistent-extras == 0.1.1",
    "python-dotenv == 1.0.1",
    "sanic-testing == 24.6.0",
    "sanic[ext] == 24.6.0",
    "setuptools == 75.1.0",
]


[project.optional-dependencies]
docgen = ["pypandoc-binary == 1.13"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest == 8.3.3",
    "ruff<1.0.0,>=0.2.2",
    "pre-commit<4.0.0,>=3.6.2",
    "coverage<8.0.0,>=7.4.3",
    "pytest-asyncio>=0.24.0",
]

[tool.ruff]
line-length = 130
target-version = "py312"

ignore = ["E731", "E701"]
# E731: we are FPers, assigning lambdas is OK
# E701: sometimes it looks better to have multiple statements after colon

[tool.ruff.per-file-ignores]
"natural4_server/*.py" = ["F403", "E701", "F405"]
# quiet warning for Joe's unused `runvue` variable
"*joe*.py" = ["F403", "E701", "F405", "E402"]
"*8090*.py" = ["F403", "E701", "F405", "E402"]
# disable false positive linting warnings for joe's config gnunicorn files
